# remove runtime recursion from find_ntuplets in Patatrack