--- a/RecoPixelVertexing/PixelTriplets/plugins/GPUCACell.h
+++ b/RecoPixelVertexing/PixelTriplets/plugins/GPUCACell.h
@@ -270,7 +270,7 @@ public:
 
   // trying to free the track building process from hardcoded layers, leaving
   // the visit of the graph based on the neighborhood connections between cells.
-  template<int DEPTH>
+  template <int DEPTH>
   __device__ inline void find_ntuplets(Hits const& hh,
                                        GPUCACell* __restrict__ cells,
                                        CellTracksVector& cellTracks,
@@ -286,7 +286,6 @@ public:
     // the ntuplets is then saved if the number of hits it contains is greater
     // than a threshold
 
-    assert(DEPTH>0);
     tmpNtuplet.push_back_unsafe(theDoubletId_);
     assert(tmpNtuplet.size() <= 4);
 
@@ -295,7 +294,7 @@ public:
       if (cells[otherCell].theDoubletId_ < 0)
         continue;  // killed by earlyFishbone
       last = false;
-      cells[otherCell].find_ntuplets<DEPTH-1>(
+      cells[otherCell].find_ntuplets<DEPTH - 1>(
           hh, cells, cellTracks, foundNtuplets, apc, quality, tmpNtuplet, minHitsPerNtuplet, startAt0);
     }
     if (last) {  // if long enough save...
@@ -348,19 +347,18 @@ private:
   hindex_type theOuterHitId;
 };
 
-
- template<>
-  __device__ inline void GPUCACell::find_ntuplets<0>(Hits const& hh,
-                                       GPUCACell* __restrict__ cells,
-                                       CellTracksVector& cellTracks,
-                                       HitContainer& foundNtuplets,
-                                       cms::cuda::AtomicPairCounter& apc,
-                                       Quality* __restrict__ quality,
-                                       TmpTuple& tmpNtuplet,
-                                       const unsigned int minHitsPerNtuplet,
-                                       bool startAt0) const {
-    assert(false);
-
-  }
+template <>
+__device__ inline void GPUCACell::find_ntuplets<0>(Hits const& hh,
+                                                   GPUCACell* __restrict__ cells,
+                                                   CellTracksVector& cellTracks,
+                                                   HitContainer& foundNtuplets,
+                                                   cms::cuda::AtomicPairCounter& apc,
+                                                   Quality* __restrict__ quality,
+                                                   TmpTuple& tmpNtuplet,
+                                                   const unsigned int minHitsPerNtuplet,
+                                                   bool startAt0) const {
+  printf("ERROR: GPUCACell::find_ntuplets reached full depth!\n");
+  assert(false);  // on GPU is compiled away in NDEBUG mode
+}
 
 #endif  // RecoPixelVertexing_PixelTriplets_plugins_GPUCACell_h