I replaced the runtime recursion with a compile time one.

compiles, runs, no regression.
is dramatically faster!
Running 4 times over 5000 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs:
before: 740.5 ±   1.4 ev/s
after: 913.9 ±  11.6 ev/s

nvprof:
before:
```
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   17.63%  2.22798s      5000  445.60us  6.5600us  1.3452ms  kernel_find_ntuplets(TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::SimpleVector<cms::cuda::VecArray<unsigned short, int=48>>*, cms::cuda::OneToManyAssoc<unsigned int, int=32769, int=163840>*, cms::cuda::AtomicPairCounter*, pixelTrack::Quality*, unsigned int)
                    9.44%  1.19250s      5000  238.50us  14.752us  1.3508ms  gpuClustering::findClus(unsigned short const *, unsigned short const *, unsigned short const *, unsigned int const *, unsigned int*, unsigned int*, int*, int)
                    9.17%  1.15818s     10000  115.82us  1.9520us  481.25us  void kernel_BLFit<int=4>(cms::cuda::OneToManyAssoc<unsigned short, int=8, int=32768> const *, double, TrackSoAHeterogeneousT<int=32768>*, double*, float*, double*, unsigned int, unsigned int)
                    8.10%  1.02397s      5000  204.79us  22.912us  1.1259ms  kernel_connect(cms::cuda::AtomicPairCounter*, cms::cuda::AtomicPairCounter*, TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::SimpleVector<cms::cuda::VecArray<unsigned int, int=36>>*, cms::cuda::VecArray<unsigned int, int=256> const *, float, float, float, float, float, float)
                    7.78%  983.37ms      5000  196.67us  10.464us  632.03us  gpuPixelDoublets::getDoubletsFromHisto(GPUCACell*, unsigned int*, cms::cuda::SimpleVector<cms::cuda::VecArray<unsigned int, int=36>>*, cms::cuda::SimpleVector<cms::cuda::VecArray<unsigned short, int=48>>*, TrackingRecHit2DSOAView const *, cms::cuda::VecArray<unsigned int, int=256>*, int, bool, bool, bool, bool, unsigned int)
                    6.89%  870.40ms      5000  174.08us  2.4000us  670.08us  void kernel_BLFit<int=3>(cms::cuda::OneToManyAssoc<unsigned short, int=8, int=32768> const *, double, TrackSoAHeterogeneousT<int=32768>*, double*, float*, double*, unsigned int, unsigned int)
                    6.48%  819.04ms      5000  163.81us  2.5600us  545.44us  gpuPixelDoublets::fishbone(TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::VecArray<unsigned int, int=256> const *, unsigned int, bool)
                    4.29%  541.47ms      5000  108.29us  2.1120us  1.2090ms  kernel_rejectDuplicate(TrackingRecHit2DSOAView const *, cms::cuda::OneToManyAssoc<unsigned int, int=32769, int=163840> const *, TrackSoAHeterogeneousT<int=32768> const *, pixelTrack::Quality*, unsigned short, bool, cms::cuda::OneToManyAssoc<unsigned short, int=-1, int=131072> const *)
                    3.81%  481.74ms      5000  96.348us  5.4070us  351.74us  gpuVertexFinder::vertexFinderOneKernel(ZVertexSoA*, gpuVertexFinder::WorkSpace*, int, float, float, float)

```
after
```
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   10.64%  993.32ms      5000  198.66us  22.656us  1.1191ms  kernel_connect(cms::cuda::AtomicPairCounter*, cms::cuda::AtomicPairCounter*, TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::SimpleVector<cms::cuda::VecArray<unsigned int, int=36>>*, cms::cuda::VecArray<unsigned int, int=256> const *, float, float, float, float, float, float)
                   10.62%  991.76ms     10000  99.176us  1.8890us  486.68us  void kernel_BLFit<int=4>(cms::cuda::OneToManyAssoc<unsigned short, int=8, int=32768> const *, double, TrackSoAHeterogeneousT<int=32768>*, double*, float*, double*, unsigned int, unsigned int)
                   10.26%  957.59ms      5000  191.52us  19.808us  1.0845ms  gpuClustering::findClus(unsigned short const *, unsigned short const *, unsigned short const *, unsigned int const *, unsigned int*, unsigned int*, int*, int)
                    9.13%  852.21ms      5000  170.44us  10.304us  690.33us  gpuPixelDoublets::getDoubletsFromHisto(GPUCACell*, unsigned int*, cms::cuda::SimpleVector<cms::cuda::VecArray<unsigned int, int=36>>*, cms::cuda::SimpleVector<cms::cuda::VecArray<unsigned short, int=48>>*, TrackingRecHit2DSOAView const *, cms::cuda::VecArray<unsigned int, int=256>*, int, bool, bool, bool, bool, unsigned int)
                    8.21%  766.72ms      5000  153.34us  2.8800us  515.87us  gpuPixelDoublets::fishbone(TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::VecArray<unsigned int, int=256> const *, unsigned int, bool)
                    8.04%  750.35ms      5000  150.07us  3.0400us  929.82us  void kernel_BLFit<int=3>(cms::cuda::OneToManyAssoc<unsigned short, int=8, int=32768> const *, double, TrackSoAHeterogeneousT<int=32768>*, double*, float*, double*, unsigned int, unsigned int)
                    5.92%  553.06ms      5000  110.61us  6.6240us  414.30us  kernel_find_ntuplets(TrackingRecHit2DSOAView const *, GPUCACell*, unsigned int const *, cms::cuda::SimpleVector<cms::cuda::VecArray<unsigned short, int=48>>*, cms::cuda::OneToManyAssoc<unsigned int, int=32769, int=163840>*, cms::cuda::AtomicPairCounter*, pixelTrack::Quality*, unsigned int)
                    4.28%  399.38ms      5000  79.875us  5.4080us  256.99us  gpuVertexFinder::vertexFinderOneKernel(ZVertexSoA*, gpuVertexFinder::WorkSpace*, int, float, float, float)

```

technical.