@cmsbuild 2021-09-29
A new Pull Request was created by  @VinInn (Vincenzo Innocente) for master.

It involves the following packages:

- RecoPixelVertexing/PixelTriplets (**reconstruction**)


@jpata, @cmsbuild, @slava77 can you please review it and eventually sign? Thanks.
@felicepantaleo, @GiacomoSguazzoni, @JanFSchulte, @rovere, @VinInn, @mmusich, @mtosi, @dgulhan this is something you requested to watch as well.
@perrotta, @dpiparo, @qliphy you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>