@VinInn 2021-09-29
For reference
```
-      thisCell.find_ntuplets<6>(
+      thisCell.find_ntuplets<16>(
```

   857.9 ±   9.7 ev/s

going to 50:

   864.9 ±   4.6 ev/s

(there is some stray activity on the machine)

down to 4 (that actually correspond to a sextuplet):

  916.0 ±   0.9 ev/s

(sextuplet are geometrically possible, not from the luminous region though)