@VinInn 2021-09-29
For those who wondered:

on CPU
Running 4 times over 5000 events with 1 jobs, each with 20 threads, 20 streams
before:  157.6 ±   3.9 ev/s
after:     158.2 ±   4.7 ev/s