@jpata 2021-09-30
assign heterogeneous