@jpata 2021-10-01
> let's put it in a issue (including test of behaviour and performance)

May I ask heterogeneous to create this issue (you know better how to describe the problem in sifficient detail)? No comments from reco, I will sign after heterogeneous has signed.