@jpata 2021-10-04
+reconstruction
- technical, no reco differences expected/observed in CPU or GPU workflows
- the error reporting in this PR can benefit from a common GPU-specific logging approach, as outlined in https://github.com/cms-sw/cmssw/issues/32502