@fwyzard 2021-10-05
For the record, here is a comparison of the throughput with respect to CMSSW_12_1_0_pre3, running over (uncompressed) TTbar events with pileup:

```
2 CPUs:
  0: Intel(R) Xeon(R) Gold 6130 CPU @ 2.10GHz (16 cores, 32 threads)
  1: Intel(R) Xeon(R) Gold 6130 CPU @ 2.10GHz (16 cores, 32 threads)

1 visible NVIDIA GPUs:
  0: Tesla T4

Running 4 times over 1000 events with 4 jobs, each with 16 threads, 16 streams and 1 GPUs
```

measurement     | CMSSW_12_1_0_pre3 |   with #35285     |   with #35473     
----------------|-------------------|-------------------|-------------------
I/O throughput  |      ~ 2 kev/s    |        ~ 2 kev/s  |        ~ 2 kev/s  
quadruplets     |   1071 ±  3 ev/s  |   1084 ±  3 ev/s  |   1289 ±  3 ev/s  
triplets        |    560 ±  2 ev/s  |    564 ±  1 ev/s  |    724 ±  3 ev/s